requirements.txt
Flask==3.0.3
openai==1.55.3
httpx==0.27.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pyvi==0.1.1
pydotenv==0.0.7
langchain-milvus==0.1.5
pymilvus[model]==2.4.7
nltk==3.9.1